WIP: split apart API tests
To help keep me more sane, in theory!
